pyproject.toml
# The MIT License (MIT)
#
# Copyright (C) 2019 - Jérémie Galarneau <jeremie.galarneau@gmail.com>
#
# Permission is hereby granted, free of charge, to any person obtaining
# a copy of this software and associated documentation files (the
# "Software"), to deal in the Software without restriction, including
# without limitation the rights to use, copy, modify, merge, publish,
# distribute, sublicense, and/or sell copies of the Software, and to
# permit persons to whom the Software is furnished to do so, subject to
# the following conditions:
#
# The above copyright notice and this permission notice shall be
# included in all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

[tool.poetry]
name = 'combotsha'
version = '0.1.0'
description = 'An IRC bot to notify of new commits'
authors = ['Jérémie Galarneau <jeremie.galarneau@gmail.com>']
repository = 'https://github.com/jgalar/combotsha/'
license = 'MIT'

[tool.poetry.dependencies]
python = '^3.6'
gitpython = '^3.1.3'
irc = '^18'

[tool.poetry.dev-dependencies]
pylint = '*'
mypy = '*'

[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'

[tool.poetry.scripts]
combotsha = 'combotsha.combotsha:_main'

[tool.black]
skip-string-normalization = true
